+++ chain/generation.py
@@ -8,13 +8,13 @@ class Generation(object):
         generate
 
         _generation:
-            generation of the text.
+            generation of a text.
 
-        _check_error:
-            check errors.
+        _check_errors:
+            check for errors.
 
         _generate_word:
-            generation of the word.
+            generation of a word.
 
         _check_punctuation:
             punctuation check.
@@ -24,20 +24,20 @@ class Generation(object):
         return self._generation(**kwargs)
 
     def _generation(self, **kwargs):
-        """Generation of the text.
+        """Generation of a text.
 
-        Generating the text based on the self._chain.
+        Generates a text based on the self._chain.
 
         Grammar:
             - the first char of a sentence will be capitalized.
-            - at the end of the text will be stand **end_char.
+            - at the end of a text will be stand **end_char.
 
         Args:
             **start (str):
                 Defaults to self.start.
-                Start of the sentence.
+                Start of a sentence.
 
-                Warning: The chain is case-sensitive 
+                Warning: the chain is case-sensitive
                 and punctuation-sensitive.
 
             **max_words (int):
@@ -59,40 +59,38 @@ class Generation(object):
 
             **end_chars (str):
                 Defaults to ".!?".
-                The chars which should stand
-                at the end of a sentence. 
+                A chars which should stand
+                at the end of a sentence.
 
             **not_end_chars (str):
                 Defaults to ','.
-                The chars which should not stand
+                A chars which should not stand
                 at the end of a sentence.
                 They will be replaced by **end_char.
 
             **end_char (str):
                 Defaults to '.'.
-                If the last char of a sentence is not in
-                **end_chars, then the **end_char 
+                If a last char of a sentence is not in
+                **end_chars, then the **end_char
                 will be appended to the end of a sentence.
 
         Returns:
             type - str.
-            If **start will found, then a text 
-            will be generated. 
-            Else the text will be equal to **start.
+            If **start will found, then a text will be generated.
+            Else a text will be equal to **start.
 
         Raises:
-            ValueError:
-                self._chain is empty.
+            see _check_errors doc.
 
         Examples:
             self._chain = {
-                'two': ['fish'], 
-                'red': ['fish'], 
-                'One': ['fish'], 
-                'fish': ['two', 'red', 'blue'], 
-                '*END*': ['*TEXT_END*'], 
-                '*START*': ['One'], 
-                'fish.': ['*END*'], 
+                'two': ['fish'],
+                'red': ['fish'],
+                'One': ['fish'],
+                'fish': ['two', 'red', 'blue'],
+                '*END*': ['*TEXT_END*'],
+                '*START*': ['One'],
+                'fish.': ['*END*'],
                 'blue': ['fish.']
             }
 
@@ -110,7 +108,7 @@ class Generation(object):
             generate(start="red", punctuation="!", end_char="!")
             Output: "Red fish two fish two fish blue fish.!"
         """
-        self._check_error()
+        self._check_errors()
 
         key = kwargs.get("start", self.start)
         max_words = kwargs.get("max_words", 20)
@@ -124,9 +122,10 @@ class Generation(object):
 
         # add a word to the end of a key.
         change_key = lambda key, word: "{} {}".format(key, word)
-        # create a key without the first word.
+        # create a key without a first word.
         create_key = lambda key: ' '.join(key.split()[1::])
 
+        # a future sentence.
         words = []
 
         # if user **start.
@@ -134,8 +133,8 @@ class Generation(object):
             words.append(key.capitalize())
         else:
             # values in self._chain[self.start]
-            # can contain self.start,
-            # self.end or self.text_end.
+            # can contain the self.start,
+            # the self.end or the self.text_end.
             word = self._generate_word(key)
             new_word = (
                 ' '.join(
@@ -193,32 +192,32 @@ class Generation(object):
 
         return ' '.join(words)
 
-    def _check_error(self):
-        """Check errors.
+    def _check_errors(self):
+        """Check for errors.
 
         Raises:
             ValueError:
-                chain is empty.
+                the self._chain is empty.
         """
         if not self._chain:
             raise ValueError("chain is empty")
 
     def _generate_word(self, key):
-        """Generation of the word.
+        """Generation of a word.
 
         Args:
-            key(str):
+            key (str):
                 a key for access to the values
                 in the self._chain.
 
         Returns:
             if key in self._chain:
                 type - str.
-                Random value.
+                The random value.
             else:
                 type - None.
         """
-        return random.choice(self._chain[key]) if key in self.chain else None
+        return random.choice(self._chain[key]) if key in self._chain else None
 
     def _check_punctuation(self, words, end_chars, not_end_chars, end_char):
         """Punctuation check.
@@ -245,6 +244,6 @@ class Generation(object):
                 words[-1] = "{}{}".format(words[-1], end_char)
             else:
                 words[-1] = "{}{}".format(
-                    words[-1][0:len(words[-1]) - 1:], 
+                    words[-1][0:len(words[-1]) - 1:],
                     end_char
                 )